rename rollup/svelte to svelte/components
