+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "dapp": "http-server web/dapp",
-    "build-dapp": "rm -rf web/dapp && npm run build:hardhat && npm run build:gulp && npm run build:rollup",
+    "build-dapp": "rm -rf web/dapp && npm run build:hardhat && npm run build:gulp && npm run build:svelte",
     "clean": "npm-run-all clean:*",
     "update": "npm-run-all update:*",
     "build": "npm-run-all build:*",
@@ -14,30 +14,30 @@
     "install:common": "cd common && pnpm install",
     "install:hardhat": "cd hardhat && pnpm install",
     "install:gulp": "cd gulp && pnpm install",
-    "install:rollup": "cd rollup && pnpm install",
+    "install:svelte": "cd svelte && pnpm install",
     "install:wordpress": "cd wordpress && pnpm install",
     "install:thegraph": "cd thegraph && pnpm install",
     "clean:common": "cd common && npm run clean",
     "clean:hardhat": "cd hardhat && npm run clean",
     "clean:gulp": "cd gulp && npm run clean",
-    "clean:rollup": "cd rollup && npm run clean",
+    "clean:svelte": "cd svelte && npm run clean",
     "clean:web": "rm -rf web/dapp web/docs",
     "clean:wordpress": "cd wordpress && npm run clean",
     "clean:thegraph": "cd thegraph && npm run clean",
     "update:common": "cd common && pnpm update",
     "update:hardhat": "cd hardhat && pnpm update",
     "update:gulp": "cd gulp && pnpm update",
-    "update:rollup": "cd rollup && pnpm update",
+    "update:svelte": "cd svelte && pnpm update",
     "update:wordpress": "cd wordpress && pnpm update",
     "update:thegraph": "cd thegraph && pnpm update",
     "build:hardhat": "cd hardhat && npm run build",
     "build:gulp": "cd gulp && npm run build",
-    "build:rollup": "cd rollup && npm run build",
+    "build:svelte": "cd svelte && npm run build",
     "build:wordpress": "cd wordpress && npm run build",
     "build:thegraph": "cd thegraph && npm run build",
     "check:common": "cd common && npm run check",
     "check:hardhat": "cd hardhat && npm run check",
-    "check:rollup": "cd rollup && npm run check",
+    "check:svelte": "cd svelte && npm run check",
     "check:wordpress": "cd wordpress && npm run check"
   },
   "repository": {
