+++ svelte/.gitignore
