+++ svelte/components/global.d.ts
