+++ svelte/components/kredeum-get-nft.svelte
@@ -14,7 +14,7 @@
     explorerNftUrl,
     explorerAddressLink
   } from "lib/kconfig";
-  import { chainId, network, owner, signer } from "./network";
+  import { chainId, network, owner, signer } from "main/network";
   import { nftGetImageLink } from "lib/knft-get";
   import { onMount } from "svelte";
 
@@ -46,10 +46,10 @@
   const divMediaVideo = (src: string, small = true) => {
     let video: string;
     if (small) {
-      video = "<video preload=\"metadata\" style=\"border-radius: initial;\">";
+      video = '<video preload="metadata" style="border-radius: initial;">';
     } else {
       video =
-        "<video autoplay=\"true\"  controls=\"\" controlslist=\"nodownload\" loop=\"\" playsinline=\"\" preload=\"metadata\" style=\"border-radius: initial;\">";
+        '<video autoplay="true"  controls="" controlslist="nodownload" loop="" playsinline="" preload="metadata" style="border-radius: initial;">';
     }
     video += `<source src="${src}" type="video/mp4"></video>`;
     return video;
@@ -71,7 +71,7 @@
     } else if (mediaType == "image") {
       div += divMediaImage(mediaSrc);
     } else {
-      div += "<div class=\"media-text\"></div>";
+      div += '<div class="media-text"></div>';
     }
     div += "</div>";
