+++ svelte/components/kredeum-nfts.svelte
@@ -11,7 +11,7 @@
   import { explorerCollectionUrl, nftsUrl, getCreate, config } from "lib/kconfig";
   import { factoryGetAddress } from "lib/kfactory-get";
 
-  import { chainId, owner } from "./network";
+  import { chainId, owner } from "main/network";
 
   export let platform = "dapp";
