+++ svelte/components/kredeum-transfer-nft.svelte
@@ -3,7 +3,7 @@
   import { transferNftResponse, transferNftReceipt } from "lib/ktransfer";
   import { explorerNftUrl, explorerTxUrl, textShort } from "lib/kconfig";
 
-  import { chainId, signer } from "./network";
+  import { chainId, signer } from "main/network";
 
   export let nft: Nft = undefined;
