+++ svelte/cypress/plugins/index.js
